feat(sdk): Rename a file
